Update FreeBSD native build for multi-target-v2.
For the bsd-kvm target (which I do not use and is not tested), I moved
several globals into the target and then changed the global kvm
commands to fetch the current process target and use a new public
method to alter the PCB address of that target.  In theory this should
permit multiple KVM targets and the commands just affect the current
inferior.

While here I cleaned up a few places to use CORE_ADDR instead of raw
pointers.  The bsd-kvm.c code is still not really ready for
cross-debugging, but this makes it slightly less terrible.

For the fbsd-nat target, I just had to sprinkle some target pointers
in various places.  There are a couple of static methods that I
added explicit target pointers as the first argument.  Alternatively
they could become private members of the class instead and just use
this.